lg-simple-reflection-agent/pyproject.toml
[project]
name = "lg-simple-reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.7",
    "langchain-community>=0.3.5",
    "langchain-openai>=0.2.6",
    "langchain-tools>=0.1.34",
    "langgraph>=0.2.45",
    "python-dotenv>=1.0.1",
]
